Nicer conversion from BasisMatrixData to matrix.
The indices of the COO sparse matrix are now calculated
in a separate function. This is almost everything that
was needed in order to generate torch matrices in a way
that preserves gradients. A further commmit will implement
this, which only requires using `torch.sparse_coo_tensor`
instead of `scipy.coo_array`, using the same indices because
indices don't need to store gradients.

If the function that calculates indices is a bottleneck, it
can be easily converted to cython in the future.