New paradigm for handling all the different formats.
Essentially, this centralizes all the handling of the formats
so that supporting conversion from everything to everything
becomes more manageable.

The initial trigger for this change was the need to support
outputing the matrices in torch format and keeping the gradients
through the process.

The main user visible change is that now `processor.matrix_from_data`
has a new `out_format` argument where you can ask for the output
to be in any supported format (including `torch_coo/csr` and `torch`
for dense tensors).

This still needs a bit of polishing over time, but it is already
quite solid. All converters are documented in a page of the
documentation.