Fixed the fill value for Hamiltonians.
For some edge matrix blocks (between different points), there can be a mix
of empty and non-empty elements (because some of the basis functions overlap
and others don't).
The empty values of the sparse matrix where until now always set to np.nan.
This meant that the model didn't care what the values produced for those matrix
elements were, which was a good approach for the density matrix.
However, the empty values of the Hamiltonian are exactly 0, and therefore the
model should learn to predict zeros for those positions. Otherwise it would predict
random numbers that would make the Hamiltonian be wrong.

In this commit, we just set the empty values automatically depending on the class
of the matrix (DensityMatrix, Hamiltonian...)