Fix MatrixWriter bug and added lightning tests.
The lightning tools were not tested and therefore it was easy that
changes in the code resulted in new bugs for them. This was a bad
situation, so here I fixed the bugs and added tests so that new
breaking changes are hopefully detected.

This is nice because it makes the lightning tools much more robust!